[svn] All times and ages are now floats, no longs anymore.

+++ src/org/codejive/world3d/TerminalEntity.java
@@ -22,7 +22,7 @@ public interface TerminalEntity {
 	 */
 	public void setLifetime(float _fLifetime);
 
-	public long getTimeOfBirth();
+	public float getTimeOfBirth();
 
 	public float getAge();
